Baru

+++ database/migrations/2024_05_15_073138_create_masyarakats_table.php
@@ -19,8 +19,8 @@ class CreateMasyarakatsTable extends Migration
             $table->string('desa');
             $table->string('nama');
             $table->integer('usia');
-            $table->double('tinggi_badan');
-            $table->double('berat_badan');
+            $table->integer('tinggi_badan');
+            $table->integer('berat_badan');
             $table->timestamps();
         });
